Feature/deploy

+++ internal/model/dto/dto.go
@@ -20,7 +20,7 @@ type BlogDto struct {
 	Category     *CategoryDto `json:"category,omitempty"`
 	Tags         []TagDto     `json:"tags,omitempty"`
 	BlogState    bool         `json:"blog_state"`
-	BlogWordsNum uint16       `json:"blog_words_num,omitempty"`
+	BlogWordsNum uint64       `json:"blog_words_num,omitempty"`
 	BlogIsTop    bool         `json:"blog_is_top"`
 	CreateTime   time.Time    `json:"create_time,omitempty"`
 	UpdateTime   time.Time    `json:"update_time,omitempty"`
